requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
crispy-bootstrap5==0.6
Deprecated==1.2.13
dj-database-url==0.5.0
dj-stripe==2.7.3
Django==4.0.5
django-celery-results==2.3.1
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-template-maths==0.2.0
gunicorn==20.1.0
idna==3.3
jsonfield==3.1.0
kombu==5.2.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.1.1
platformdirs==3.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.3
requests==2.28.0
sib-api-v3-sdk==7.5.0
six==1.16.0
sqlparse==0.4.2
stripe==4.2.0
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
